#16: also update UNIX offset constant to larger data range..

+++ adma_ros2_driver/src/adma_driver.cpp
@@ -121,7 +121,7 @@ void ADMADriver::parseData(std::array<char, 856> recv_buf)
   message_imu.header.frame_id = imu_frame_;
   float weektime;
   //offset between UNIX and GNSS (in ms)
-  unsigned long offset_gps_unix = 315964800000;
+  unsigned long long offset_gps_unix = 315964800000;
   uint32_t week_to_msec = 604800000;
   unsigned long long timestamp;
